backend/requirements.txt
opencv-python
mediapipe
pymongo
python-dotenv
fastapi
uvicorn
python-multipart
websockets